requirements.txt
openai-agents==0.0.3
openai>=1.66.2
python-dotenv==1.0.1
flask==3.0.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.123.0
icalendar==5.0.11
vobject==0.9.6.1
pytz==2024.1 